+++ inc/home/homepage-artist-card-actions.php
@@ -35,7 +35,7 @@ function ec_add_artist_card_management_buttons( $artist_id ) {
     }
 
     // Get management URLs
-    $manage_artist_page = get_page_by_path( 'manage-artist-profiles' );
+    $manage_artist_page = get_page_by_path( 'manage-artist' );
     $manage_link_page = get_page_by_path( 'manage-link-page' );
 
     $manage_artist_url = $manage_artist_page ? get_permalink( $manage_artist_page ) : '';